main.py/requirements.txt

fastapi
uvicorn
gspread
oauth2client
python-telegram-bot==13.15
